requirements.txt
addict>=2.2.0
fire>=0.1.3
PyYAML>=3.13
jinja2>=2.10
toolz>=0.9.0